lightly wrap as test before building integration tests
